+++ src/Helpers.php
@@ -3,9 +3,12 @@
 namespace Nearata\MinecraftAvatars;
 
 use Illuminate\Http\Client\Factory;
+use Illuminate\Support\Arr;
 
 class Helpers
 {
+    public static $usernames = ['MHF_Steve', 'MHF_Alex'];
+
     public static function isUsername(string $minotar): bool
     {
         $validUsernameRegex = '/^[a-zA-Z0-9_]{1,16}$/';
@@ -21,5 +24,19 @@ class Helpers
         }
 
         return $response->json('id');
+    }
+
+    public static function getRandomUsername(): string
+    {
+        return Arr::random(Helpers::$usernames);
+    }
+
+    public static function isUUID(string $minotar): bool
+    {
+        $validUUIDPlainRegex = '[0-9a-f]{32}';
+        $validUUIDDashRegex = '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}';
+        $validUUIDRegex = '/^(' . $validUUIDPlainRegex . '|' . $validUUIDDashRegex . ')$/';
+
+        return preg_match($validUUIDRegex, $minotar);
     }
 }